+++ src/libultra/io/sptask.c
@@ -32,29 +32,23 @@ void osSpTaskLoad(OSTask* intp) {
         intp->t.flags &= ~OS_TASK_YIELDED;
 
         if (tp->t.flags & OS_TASK_LOADABLE) {
-            tp->t.ucode = HW_REG((u32)intp->t.yieldDataPtr + OS_YIELD_DATA_SIZE - 4, u32);
+            tp->t.ucode = HW_REG((uintptr_t)intp->t.yieldDataPtr + OS_YIELD_DATA_SIZE - 4, u32);
         }
     }
     osWritebackDCache(tp, sizeof(OSTask));
     __osSpSetStatus(SP_CLR_SIG0 | SP_CLR_SIG1 | SP_CLR_SIG2 | SP_SET_INTR_BREAK);
 
-    while (__osSpSetPc((void*)SP_IMEM_START) == -1) {
-        ;
-    }
-    while (__osSpRawStartDma(1, (void*)(SP_IMEM_START - sizeof(*tp)), tp, sizeof(OSTask)) == -1) {
-        ;
-    }
-    while (__osSpDeviceBusy()) {
-        ;
-    }
-    while (__osSpRawStartDma(1, (void*)SP_IMEM_START, tp->t.ucodeBoot, tp->t.ucodeBootSize) == -1) {
-        ;
-    }
+    while (__osSpSetPc((void*)SP_IMEM_START) == -1) {}
+
+    while (__osSpRawStartDma(1, (void*)(SP_IMEM_START - sizeof(*tp)), tp, sizeof(OSTask)) == -1) {}
+
+    while (__osSpDeviceBusy()) {}
+
+    while (__osSpRawStartDma(1, (void*)SP_IMEM_START, tp->t.ucodeBoot, tp->t.ucodeBootSize) == -1) {}
 }
 
 void osSpTaskStartGo(OSTask* tp) {
-    while (__osSpDeviceBusy()) {
-        ;
-    }
+    while (__osSpDeviceBusy()) {}
+
     __osSpSetStatus(SP_SET_INTR_BREAK | SP_CLR_SSTEP | SP_CLR_BROKE | SP_CLR_HALT);
 }
